Add created_at field to all ddb entries
